% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cai1.R
\name{cai1}
\alias{cai1}
\title{Calculate Chloro-Alkaline Index 1 (CAI-1)}
\usage{
cai1(data, base_unit = c("mg/L", "mmol/L", "meq/L"))
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table} containing the required ion
columns \code{"Na"}, \code{"K"}, and \code{"Cl"}.}

\item{base_unit}{Unit of the input ion concentrations. One of \code{"mg/L"},
\code{"mmol/L"}, or \code{"meq/L"}. Default is \code{"mg/L"}.}
}
\value{
A numeric vector of CAI-1 values, one per row of \code{data}.
}
\description{
Computes the Chloro-Alkaline Index 1 (CAI-1) from hydrochemical data using
sodium, potassium, and chloride concentrations. Input concentrations may be
provided in \code{"mg/L"}, \code{"mmol/L"}, or \code{"meq/L"} and are internally converted
to \code{"meq/L"} prior to calculation.
}
\details{
The function checks that all required ions (\code{Na}, \code{K}, \code{Cl}) are present in
the input dataset and throws an error if any are missing.
}
\examples{
data("hc_data", package = "hydrochem")
cai1(hc_data, base_unit = "mg/L")

}
