% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cai2.R
\name{cai2}
\alias{cai2}
\title{Calculate Chloro-Alkaline Index 2 (CAI-2)}
\usage{
cai2(data, base_unit = c("mg/L", "mmol/L", "meq/L"))
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table} containing the required ion
columns \code{"Na"}, \code{"K"}, \code{"Cl"}, \code{"SO4"}, \code{"HCO3"}, and \code{"NO3"}.}

\item{base_unit}{Unit of the input ion concentrations. One of \code{"mg/L"},
\code{"mmol/L"}, or \code{"meq/L"}. Default is \code{"mg/L"}.}
}
\value{
A numeric vector of CAI-2 values, one per row of \code{data}.
}
\description{
Computes the Chloro-Alkaline Index 2 (CAI-2) from hydrochemical data using
sodium, potassium, chloride, sulfate, bicarbonate, nitrate, and carbonate
concentrations. Input concentrations may be provided in \code{"mg/L"}, \code{"mmol/L"},
or \code{"meq/L"} and are internally converted to \code{"meq/L"} prior to calculation.
}
\details{
The function checks that all required ions (\code{Na}, \code{K}, \code{Cl}, \code{SO4}, \code{HCO3}, \code{NO3}) are present in the input dataset and throws an error if any are
missing. If carbonate (\code{CO3}) is also provided, its contribution is added
to bicarbonate after conversion to milliequivalents (\code{meq/L}). If \code{CO3} is absent, alkalinity is considered to be represented solely by bicarbonate (\code{HCO3}).
}
\examples{
data("hc_data", package = "hydrochem")
cai2(hc_data, base_unit = "mg/L")

}
