% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_unit.R
\name{convert_unit}
\alias{convert_unit}
\title{Convert Ion Concentrations Between Units (mg/L, mmol/L, meq/L)}
\usage{
convert_unit(
  data,
  ions = c("Ca", "Mg", "Na", "K", "Cl", "SO4", "HCO3"),
  base_unit = c("mg/L", "mmol/L", "meq/L"),
  to_unit = c("mg/L", "mmol/L", "meq/L"),
  convert_to_species = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table} with numeric columns named after ions (e.g., \code{"Ca"}, \code{"Cl"}).}

\item{ions}{A character vector of ions names. Default is \code{c("Ca", "Mg", "Na", "K", "Cl", "SO4", "HCO3")}.}

\item{base_unit}{Unit of the input ion concentrations. One of \code{"mg/L"}, \code{"mmol/L"}, or \code{"meq/L"}. Default is \code{"mg/L"}.}

\item{to_unit}{Unit of the output unit. One of \code{"mg/L"}, \code{"mmol/L"}, or \code{"meq/L"}. Default is \code{"mg/L"}.}

\item{convert_to_species}{Logical. If \code{TRUE}, converts known elemental forms (e.g., \code{"NO3_N"}, \code{"PO4_P"}, \code{"H2S_HS_S"}) to their compound species (\code{"NO3"}, \code{"PO4"}, \code{"H2S"}).}
}
\value{
A \code{data.table} with original columns and new columns suffixed based on the \code{to_unit} function parameter.
}
\description{
Converts concentrations of selected ions (cations and anions) between supported units:
\code{"mg/L"}, \code{"mmol/L"}, and \code{"meq/L"}. Input columns must be named using
plain ion names (e.g., \code{"Ca"}, \code{"Cl"}). Output columns are added with suffixes:
\verb{_mg}, \verb{_mmol}, and \verb{_meq} depending on the \code{to_unit} function parameter.
}
\examples{
data("hc_data", package = "hydrochem")
convert_unit(hc_data, base_unit = "mg/L", to_unit = "meq/L")

}
