% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_qc.R
\docType{data}
\name{data_qc}
\alias{data_qc}
\title{Hydrochemical dataset from Quebec}
\format{
A data frame with 146 rows of hydrochemical data.
}
\source{
Cloutier V. (2004) Origin and goechemical evolution of groundwater in the Paleozoic Basses-Laurentides sedimentary rock aquifer system, St. Lawrence Lowlands, Québec, Canada. PhD Thesis, INRS-Eau, Terre & Environnement, Québec, Canada (electronic version on \url{https://espace.inrs.ca/id/eprint/1445/})
}
\usage{
data_qc
}
\description{
A dataset containing ion concentrations in mg/L for applying hydrochemical tools.
}
\examples{
data(data_qc)
head(data_qc)

}
\keyword{datasets}
