% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_data.R
\docType{data}
\name{hc_data}
\alias{hc_data}
\title{Example hydrochemical dataset}
\format{
A data frame with 59 rows of hydrochemical data.
}
\source{
Anonymized field data
}
\usage{
hc_data
}
\description{
A sample dataset containing ion concentrations in mg/L for applying hydrochemical tools.
}
\examples{
data(hc_data)
head(hc_data)

}
\keyword{datasets}
