% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnesium_hazard.R
\name{magnesium_hazard}
\alias{magnesium_hazard}
\title{Calculate Magnesium Hazard (MH)}
\usage{
magnesium_hazard(data, base_unit = c("mg/L", "mmol/L", "meq/L"))
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table} containing the required ion
columns \code{"Ca"} and \code{"Mg"}.}

\item{base_unit}{Unit of the input ion concentrations. One of \code{"mg/L"},
\code{"mmol/L"}, or \code{"meq/L"}. Default is \code{"mg/L"}.}
}
\value{
A numeric vector of Magnesium Hazard values (in \%), one per row of
\code{data}.
}
\description{
Computes the Magnesium Hazard (MH) index from hydrochemical data using
calcium and magnesium concentrations. Input concentrations may be provided in
\code{"mg/L"}, \code{"mmol/L"}, or \code{"meq/L"} and are internally converted to \code{"meq/L"}
prior to calculation.
}
\details{
The function checks that both required ions (\code{Ca}, \code{Mg}) are present in the
input dataset and throws an error if either is missing.
}
\examples{
data("hc_data", package = "hydrochem")
magnesium_hazard(hc_data, base_unit = "mg/L")

}
