% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_sodium.R
\name{percent_sodium}
\alias{percent_sodium}
\title{Calculate Percent Sodium (\%Na)}
\usage{
percent_sodium(data, base_unit = c("mg/L", "mmol/L", "meq/L"))
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table} containing the required ion
columns \code{"Na"}, \code{"K"}, \code{"Ca"}, and \code{"Mg"}.}

\item{base_unit}{Unit of the input ion concentrations. One of \code{"mg/L"},
\code{"mmol/L"}, or \code{"meq/L"}. Default is \code{"mg/L"}.}
}
\value{
A numeric vector of Percent Sodium values (in \%), one per row of \code{data}.
}
\description{
Computes the Percent Sodium (\%Na) index from hydrochemical data using sodium,
potassium, calcium, and magnesium concentrations. Input values may be
provided in \code{"mg/L"}, \code{"mmol/L"}, or \code{"meq/L"} and are internally converted
to \code{"meq/L"} prior to calculation.
}
\details{
The function checks that all required ions (\code{Na}, \code{K}, \code{Ca}, \code{Mg}) are present
in the input dataset and throws an error if any are missing.
}
\examples{
data("hc_data", package = "hydrochem")
percent_sodium(hc_data, base_unit = "mg/L")

}
