% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permeability_index.R
\name{permeability_index}
\alias{permeability_index}
\title{Calculate Permeability Index (PI)}
\usage{
permeability_index(data, base_unit = c("mg/L", "mmol/L", "meq/L"))
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table} containing the required ion
columns \code{"Na"}, \code{"HCO3"}, \code{"Ca"}, and \code{"Mg"}.}

\item{base_unit}{Unit of the input ion concentrations. One of \code{"mg/L"},
\code{"mmol/L"}, or \code{"meq/L"}. Default is \code{"mg/L"}.}
}
\value{
A numeric vector of Permeability Index values (in \%), one per row of
\code{data}.
}
\description{
Computes the Permeability Index (PI) from hydrochemical data using sodium,
bicarbonate, calcium, and magnesium concentrations. Input concentrations may
be provided in \code{"mg/L"}, \code{"mmol/L"}, or \code{"meq/L"} and are internally converted to
\code{"meq/L"} prior to calculation.
}
\details{
The function checks that all required ions (\code{Na}, \code{HCO3}, \code{Ca}, \code{Mg}) are
present in the dataset and throws an error if any are missing.
}
\examples{
data("hc_data", package = "hydrochem")
permeability_index(hc_data, base_unit = "mg/L")

}
