% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save a plot Diagram to File}
\usage{
save_plot(
  filename,
  type = NULL,
  plot = last_plot(),
  dpi = 300,
  width = NULL,
  height = NULL,
  ...
)
}
\arguments{
\item{filename}{Character. Path to the output file (e.g., \code{"piper_plot.png"}).}

\item{type}{Character. Diagram type used to choose default width/height if not
supplied. One of \verb{c("piper", "ternary", "collins", "stabler", "durov", "gibbs", schoeller", "stiff", "biplot")}.}

\item{plot}{A ggplot object to save. Default is the last plot (\code{last_plot()}).}

\item{dpi}{Resolution of the output image in dots per inch. Default is 300.}

\item{width, height}{Numeric. Output size. If \code{NULL}, defaults depend on \code{type}.}

\item{...}{Additional arguments passed to \code{ggsave()}.}
}
\value{
Invisibly returns the filename.
}
\description{
Saves a plot diagram (or any ggplot object) to a file using \code{ggsave()}.
}
\examples{
p <- plot_piper(hc_data)
\donttest{
  save_plot(file.path(tempdir(), "piper.png"), type = "piper", plot = p)
}

p <- plot_stiff(hc_data[1:10, ], ncol = 5)
\donttest{
  if (interactive()) {
    save_plot(file.path(tempdir(), "stiff.png"), type = "stiff", plot = p, width = 10, height = 4)
  }
}
}
