% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_summary.R
\name{stat_summary}
\alias{stat_summary}
\title{Summary Statistics}
\usage{
stat_summary(data, params = "all", group = NULL, stats = NULL, na.rm = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table} with numeric columns.}

\item{params}{A character vector of cols names to summarize. If \code{params = "all"}, all numerical columns are used.}

\item{group}{Optional character string. Name of the grouping column in \code{data}. Default is NULL.}

\item{stats}{Character vector of statistics to select (e.g., \code{c("min", "max", "mean")}).
Must be among: \code{c("min", "P1", "P5", "P10", "P25", "med", "P75", "P90", "P95", "P99",
  "max", "mean", "gmean", "mode", "range", "IQR", "sd", "mad", "CV", "n", "n_percent",
  "NA", "NA_percent")}. If \code{NULL}, all available statistics are returned.}

\item{na.rm}{Logical. Should missing values be removed before computing statistics? Default is \code{FALSE}.}
}
\value{
A \code{data.table} with summarized statistics for the selected columns.
}
\description{
Computes summary statistics for selected numeric columns in a dataset, including:
mean, standard deviation, minimum, maximum, median, percentiles (1\%, 5\%, 25\%, 75\%, 95\%, 99\%),
geometric mean, mode, median absolut deviation (MAD), coefficient of variation (CV), range, interquartile range (IQR),
number of non-missing values (n), percentage of non-missing values (n_percent), number of NAs (NA), and percentage  of NAs (NA_percent).
}
\examples{
data("hc_data", package = "hydrochem")
stat_summary(hc_data, params = c("Ca", "Mg", "Na"))
stat_summary(hc_data, group = "type", params = c("Ca", "Mg", "Na")) 

}
